apollo_defconfig: disable ZRAM
while it may seem that I just steal from proton (maybe)
this took much discussion and testing to come to a conclusion to disable it.

from my testing, ZRAM is not really needed due to the absurd amount of RAM that we have.
though, if you think otherwise, criticism is more than welcome.

Signed-off-by: Lucas Lee Jing Yi <lucasleeeeeeeee@gmail.com>